x86: Add explicit type conversion for encode/decode pointer
Eliminate warning on 32bit builds.

+++ arch/x86/include/asm/types.h
@@ -126,7 +126,7 @@ typedef u32 tls_t;
 
 typedef UserX86RegsEntry UserRegsEntry;
 
-static inline u64 encode_pointer(void *p) { return (u64)p; }
-static inline void *decode_pointer(u64 v) { return (void*)v; }
+static inline u64 encode_pointer(void *p) { return (u64)(long)p; }
+static inline void *decode_pointer(u64 v) { return (void*)(long)v; }
 
 #endif /* __CR_ASM_TYPES_H__ */
